Added functionality for "for" loops.
It does not give complexity line by line yet but it can give complexity for a few basic for statements. It cannot give correct complexity for statements like:
int i=10,j;
for(j=1;j<=i;j++)
My code will show it to be O(n) while the actual is O(1). The changes are yet to be done.